gcs to pubsub cf/requirements.txt
google-cloud-pubsub
google-cloud-storage
google-cloud-core
pytz